backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
numpy
python-dotenv
pytest